+++ RELEASENOTES.md
@@ -12,6 +12,9 @@
     *   Fix potential `IndexOutOfBoundsException` caused by extractors reporting
         additional tracks after the initial preparation step
         ([#1476](https://github.com/androidx/media/issues/1476)).
+    *   `Effects` in `ExoPlayer.setVideoEffect()` will receive the timestamps
+        with the renderer offset removed
+        ([#1098](https://github.com/androidx/media/issues/1098)).
 *   Transformer:
 *   Track Selection:
 *   Extractors: